requirements.txt

# Additional requirements go here
django-debug-toolbar<4.1.0
